+++ packages/react-ui/src/index.js
@@ -8,6 +8,7 @@ import LegendWidgetUI, { LEGEND_TYPES } from './widgets/legend/LegendWidgetUI';
 import ScatterPlotWidgetUI from './widgets/ScatterPlotWidgetUI';
 import TimeSeriesWidgetUI from './widgets/TimeSeriesWidgetUI/TimeSeriesWidgetUI';
 import { CHART_TYPES } from './widgets/TimeSeriesWidgetUI/utils/constants';
+import NoDataAlert from './widgets/NoDataAlert';
 
 export {
   cartoThemeOptions,
@@ -20,5 +21,6 @@ export {
   TimeSeriesWidgetUI,
   CHART_TYPES as TIME_SERIES_CHART_TYPES,
   LegendWidgetUI,
-  LEGEND_TYPES
+  LEGEND_TYPES,
+  NoDataAlert
 };