+++ packages/react-widgets/src/index.d.ts
@@ -14,4 +14,3 @@ export {
 } from './models';
 export { default as TimeSeriesWidget } from './widgets/TimeSeriesWidget';
 export { default as useSourceFilters } from './hooks/useSourceFilters';
-export { default as NoDataAlert } from './widgets/NoDataAlert';