+++ packages/react-widgets/src/index.js
@@ -14,4 +14,3 @@ export {
   getScatter
 } from './models';
 export { default as useSourceFilters } from './hooks/useSourceFilters';
-export { default as NoDataAlert } from './widgets/NoDataAlert';