+++ packages/react-widgets/src/widgets/ScatterPlotWidget.js
@@ -2,10 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { PropTypes } from 'prop-types';
 import { selectIsViewportFeaturesReadyForSource } from '@carto/react-redux';
-import { WrapperWidgetUI, ScatterPlotWidgetUI } from '@carto/react-ui';
+import { WrapperWidgetUI, ScatterPlotWidgetUI, NoDataAlert } from '@carto/react-ui';
 import { getScatter } from '../models';
 import useSourceFilters from '../hooks/useSourceFilters';
-import NoDataAlert from './NoDataAlert';
 
 /**
  * Renders a <ScatterPlotWidget /> component